Add DSM binding function in hardwired format.
This patch add the needed train pulse to put a Spektrum Satellite
in bind mode. The function is hardcoded and then a specific boot
loader is needed to do this. A better solution could be to drive
Satellite power line through. This GPIO could switch off and then
ON the Satellite before binding pulse generation.

Signed-off-by: Giuseppe Barba <giuseppe.barba@st.com>